+++ smartsheet/models/__init__.py
@@ -121,3 +121,4 @@ from .webhook_subscope import WebhookSubscope
 from .widget import Widget
 from .widget_content import WidgetContent
 from .workspace import Workspace
+from .user_plan import UserPlan